fix(welllog): green C++ CI for -Werror, shared link, packaging, Windows
Unblock WellLogEngine workflow after harfbuzz deps landed:

- Clang -Werror: uint16 LE cast, unused lambda capture, LIS f16 sign
- Export detail raster APIs so shared libwelllog-render-gl links tests
- find_dependency(ZLIB) for package.consumer with static IO/export
- Soften append-coalescing race; grow PDF inflate buffer for band counts
- Qt tests: ignore -Wconversion on QTRY_* macros
- Windows: MSVC + vcpkg zlib; wheel Qt/PySide 6.8.3 (aqt lacks 6.11.1)

+++ well-log-engine/src/io/container_security.cpp
@@ -24,8 +24,11 @@ namespace {
 }
 
 [[nodiscard]] std::uint16_t read_u16_le(const std::byte *p) {
-  return static_cast<std::uint16_t>(static_cast<std::uint8_t>(p[0])) |
-         (static_cast<std::uint16_t>(static_cast<std::uint8_t>(p[1])) << 8);
+  // Cast the full expression: the `|` of two promoted ints is an `int` under
+  // usual arithmetic conversions; Clang -Wimplicit-int-conversion flags that.
+  return static_cast<std::uint16_t>(
+      static_cast<std::uint16_t>(static_cast<std::uint8_t>(p[0])) |
+      (static_cast<std::uint16_t>(static_cast<std::uint8_t>(p[1])) << 8));
 }
 
 [[nodiscard]] std::uint32_t read_u32_le(const std::byte *p) {
